Also test rawstatus/purge file job
